pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "os-explorer"
version = "0.1.0"
description = "OpenStack Resource Explorer and Graph Visualizer"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openstacksdk>=1.0.0",
    "rich>=12.0.0",
    "typer>=0.7.0",
    "textual>=0.20.0",
    "networkx>=3.0"
]

[project.scripts]
os-explorer = "os_explorer.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
